% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_bearing_utils.R
\name{angle_is_between}
\alias{angle_is_between}
\title{Test if one or more angles are between two reference angles}
\usage{
angle_is_between(x, a, b)
}
\arguments{
\item{x}{A vector of one or more angles to test.}

\item{a}{The first reference angle (start of arc).}

\item{b}{The second reference angle (end of arc).}
}
\value{
A logical vector the same length as x.
}
\description{
Given references angles a and b, this function checks if the angle(s) x are
between a and b. This will be true when a given angle lies in the clockwise
arc starting at a and ending at b. Adapted from an example posted by Jonathan
Mee on Stack Overflow: https://stackoverflow.com/a/42424631/40246.
}
