% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_bearing_utils.R
\name{compass_is_clockwise}
\alias{compass_is_clockwise}
\title{Check if the shortest rotation two bearings is clockwise}
\usage{
compass_is_clockwise(b1, b2)
}
\arguments{
\item{b1}{First compass bearing (degrees).}

\item{b2}{Second compass bearing (degrees).}
}
\value{
A list with two named elements: clockwise (logical) and angle (shortest
angular rotation between the two bearings).
}
\description{
Given two compass bearings, this function checks if the shortest rotation
from the first to the second is clockwise. For cases where the two bearings
are exactly opposite, e.g. 90 (E) and 270 (W), the function returns TRUE.
Adapted from a post by Allan Cameron on StackOverflow:
https://stackoverflow.com/a/63429855/40246.
}
\examples{
# NW to NE
compass_is_clockwise(315, 45)  # returns list(TRUE, 45)

# NE to NW
compass_is_clockwise(45, 315)  # returns list(FALSE, 45)

# Exactly opposite bearings are treated as clockwise
compass_is_clockwise(270, 90)  # returns list(TRUE, 180)

}
