% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_bearing_utils.R
\name{compass_to_rad}
\alias{compass_to_rad}
\title{Convert compass bearings in degrees to Cartesian angles in radians}
\usage{
compass_to_rad(x)
}
\arguments{
\item{x}{Vector of one or more compass bearings in degrees.}
}
\value{
Vector of corresponding Cartesian angles in radians in the range
[0, 2*pi).
}
\description{
Cartesian angles, as used by R's trigonometric functions, treat 0 as the
positive X-axis direction, with positive angles being counter-clockwise
rotation. In contrast, compass angles treat 0 as North with positive angles
being clockwise rotation.
}
