% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_focal_matrix.R
\name{make_focal_matrix}
\alias{make_focal_matrix}
\title{Create a weights matrix for use as a circular or sector focal neighbourhood}
\usage{
make_focal_matrix(
  radius,
  start_bearing = NULL,
  end_bearing = NULL,
  include_centre = TRUE,
  inside_value = 1,
  outside_value = NA
)
}
\arguments{
\item{radius}{(integer > 0) Radius of the neighbourhood (pixels).}

\item{start_bearing}{Optional compass bearing (degrees) to define a sector.
Default value @code{NULL} means create a full circular neighbourhood.}

\item{end_bearing}{Optional compass bearing (degrees) to define a sector.
Ignored if \code{start_bearing} is NULL.}

\item{include_centre}{(logical, default TRUE) Whether to include the centre
cell in the neighbourhood.}

\item{inside_value}{(default 1) Value to assign to cells inside the
neighbourhood.}

\item{outside_value}{(default NA) Value to assign to cells outside the
neighbourhood.}
}
\description{
Creates a binary weights matrix representing a circular neighbourhood to use
with the \code{\link[terra]{focal}} function. Note, if you are specifying any
function other than \code{"mean"} with the focal function you should adjust
the weight values accordingly. Optionally, a sector neighbourhood can be
constructed rather than a full circular neighbourhood. The sector is defined
via the \code{start_bearing} and \code{end_bearing} arguments. The arc of
the sector will go clockwise from the start to the end bearing.
}
\examples{
# Create a neighbourood with a radius of 10 pixels. Returns a 21 x 21 matrix.
w <- make_focal_matrix(10)

# Create a sector neighbourhood facing West with angular width of 90 degrees
# (i.e. SW to NW directions).
#
w <- make_focal_matrix(10, 225, 315)

# Create a sector neighbourhood that excludes pixels between SW and NW
# directions.
#
w <- make_focal_matrix(10, 315, 225)

}
