% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_bearing_utils.R
\name{normalize_radians}
\alias{normalize_radians}
\title{Put Cartesian angles into the range [0, 2*pi)}
\usage{
normalize_radians(x)
}
\arguments{
\item{x}{Vector of one or more Catesian angles (radians).}
}
\value{
Corresponding values in the range [0, 2*pi).
}
\description{
Put Cartesian angles into the range [0, 2*pi)
}
\examples{
normalize_radians(-pi/4) # returns 7*pi/4 (approx 5.4978)

}
