% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_bearing_utils.R
\name{rad_to_compass}
\alias{rad_to_compass}
\title{Convert Cartesian angles in radians to compass bearings in degrees}
\usage{
rad_to_compass(x)
}
\arguments{
\item{x}{Vector of one or more Cartesian angles in radians.}
}
\value{
Vector of corresponding compass angles in the range [0,360).
}
\description{
Cartesian angles, as used by R's trigonometric functions, treat 0 as the
positive X-axis direction, with positive angles being counter-clockwise
rotation. In contrast, compass angles treat 0 as North with positive angles
being clockwise rotation.
}
